feat(listbox): move controls to the utils folder
